requirement.txt
huggingface_hub
pillow
python-multipart
fastapi
requests
